feat: split once primitive features
