+++ asyncband/src/broadcast/overflow/mod.rs
@@ -73,10 +73,10 @@ use std::sync::atomic::Ordering;
 use std::task::Context;
 use std::task::Poll;
 
-use crate::internal::Mutex;
-use crate::internal::RwLock;
-use crate::internal::WaitRegistration;
-use crate::internal::WaitSet;
+use crate::internal::mutex::Mutex;
+use crate::internal::rwlock::RwLock;
+use crate::internal::waitset::WaitRegistration;
+use crate::internal::waitset::WaitSet;
 
 #[cfg(test)]
 mod tests;
