+++ asyncband/src/latch/mod.rs
@@ -62,8 +62,8 @@ use std::sync::Arc;
 use std::task::Context;
 use std::task::Poll;
 
-use crate::internal::CountdownState;
-use crate::internal::WaitRegistration;
+use crate::internal::countdown::CountdownState;
+use crate::internal::waitset::WaitRegistration;
 
 /// A synchronization primitive that can be used to coordinate multiple tasks.
 ///
