+++ asyncband/src/once/once/mod.rs
@@ -20,8 +20,8 @@ use std::pin::Pin;
 use std::task::Context;
 use std::task::Poll;
 
-use crate::internal::CountdownState;
-use crate::internal::WaitRegistration;
+use crate::internal::countdown::CountdownState;
+use crate::internal::waitset::WaitRegistration;
 use crate::semaphore::Semaphore;
 
 /// A synchronization primitive which can be used to run a one-time async initialization.
